# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "3.0.0",
    "title": "Handwriter Express Application",
    "description": "Handwriter Express Application Routes"
  },
  "host": "handwritingbackend.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/email/trigger/otp": {
      "post": {
        "tags": [
          "Email"
        ],
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "body",
            "description": "Email id required for OTP",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "200shreyans@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP sent to user",
            "schema": {
              "type": "object",
              "properties": {
                "verification_key": {
                  "type": "string",
                  "example": "U2FsdGVkX1+ZMdUr8NfZX/FCeKFc1fdrnsVm1AONDODv6kxbHuQ/8ZoQcrAN2zmxUZUBWFn6ihnpHW+s/gNpc+Mg4pxk++JoMix1N3stdI4aAYSKdYsPx+V/tqvSgOGijabkUW1kRLcwWeMlTJO/2Q=="
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Email not provided.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Email not provided"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error occured"
          }
        }
      }
    },
    "/email/verify/otp": {
      "post": {
        "tags": [
          "Email"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "OTP Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "verification_key": {
                  "type": "string",
                  "example": "Verification Key from trigger OTP"
                },
                "otp": {
                  "type": "string",
                  "example": "XXXXXX"
                },
                "check": {
                  "type": "string",
                  "example": "Email ID to be verified"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP verified by user",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "SUCCESS"
                },
                "message": {
                  "type": "string",
                  "example": "OTP Matched"
                },
                "check": {
                  "type": "string",
                  "example": "200shreyans@gmail.com"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Multiple 400 errors",
            "schema": {
              "type": "object",
              "properties": {
                "VerificationKey_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Verification_key not provided"
                    }
                  }
                },
                "OTP_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "OTP not provided"
                    }
                  }
                },
                "WrongKey_Error_or_CorouptedKey_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Verification key corupted"
                    }
                  }
                },
                "Check_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Email ID Incorrect"
                    }
                  }
                },
                "WrongOtp_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "OTP NOT Matched"
                    }
                  }
                },
                "OTPExpired_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "OTP Expired"
                    }
                  }
                },
                "OTPAlreadyUsed_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "OTP already used"
                    }
                  }
                },
                "BadRequest_Error": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error occured"
          }
        }
      }
    },
...
    "/user/category/create/category": {
      "post": {
        "tags": [
          "User Category"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Create Category",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "categoryName": {
                  "type": "string",
                  "example": "INTRO"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User Cateories",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "example": "61c8af95598a5a9b7a56ee92"
                  },
                  "title": {
                    "type": "string",
                    "example": "INTRO"
                  },
                  "dialogues": {
                    "type": "array",
                    "example": [],
                    "items": {}
                  },
                  "__v": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "User not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bad Request"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Token BlackListed"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error occured"
          }
        }
      }
    },
    "/user/category/update/dialogue": {
      "post": {
        "tags": [
          "User Category"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Update user dialogue",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "categoryId": {
                  "type": "string",
                  "example": "61c8af95598a5a9b7a56ee92"
                },
                "dialogue": {
                  "type": "string",
                  "example": "abcd"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User Cateories",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string",
                    "example": "61c8af95598a5a9b7a56ee92"
                  },
                  "title": {
                    "type": "string",
                    "example": "INTRO"
                  },
                  "dialogues": {
                    "type": "array",
                    "example": [
                      "abcd"
                    ],
                    "items": {
                      "type": "string"
                    }
                  },
                  "__v": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "User not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bad Request"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Token BlackListed"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error occured"
          }
        }
      }
    }
  },
...
}